add initializator for constructor with driver, needs for plugin
