+++ src/main/resources/META-INF/defaultservices/com.codeborne.selenide.Clipboard
@@ -1,1 +1,1 @@
-com.codeborne.selenide.ClipboardLocal
+com.codeborne.selenide.Clipboard
